Added ACME emulators device folder
The ACME emulators device folder 'device/acme/generic' is now added on
top of the official AOSP source tree.

The ACME emulators `acme_x86_64` and `acme_car_x86_64`, for both
classic Phone & Tablet Android and Android Automotive, will be used as
base for adding and/or replacing components and features.